Feat: Animation when changing pages
I swear to god I cannot figure out for the life of me how to fix the pong-menu background animation. I don’t even know how to put it into words, but I’ll just leave it like this for now (forever, I’m not touching this ever again).

Ahem, this commit adds these animations to both pong.html and pong-menu.html. It works by passing the parameter `animate=true` as a GET request parameter and deletes this param on load of the page.